Dynamic resource handling on a per object basis
